improve logging, and parse some statistics

+++ smj-ulm-cal/admin/class-smj-ulm-cal-admin.php
@@ -248,9 +248,13 @@ function smj_ulm_cal_setting__num_sync_calendars() {
 function smj_ulm_cal_delete_cache() {
 	if ( is_user_logged_in() ) {
 		$dir_path = plugin_dir_path(__FILE__) ."../data/calender.ics";
-		unlink($dir_path);
-		$dir_path = plugin_dir_path(__FILE__) ."../data/categories.txt";
-		unlink($dir_path);
+		if(is_file($dir_path)) {
+			unlink($dir_path);
+		}
+		$dir_path = plugin_dir_path(__FILE__) ."../data/statistic.txt";
+		if(is_file($dir_path)) {
+			unlink($dir_path);
+		}
 		wp_redirect(admin_url("admin.php?page=smj_ulm_cal_options__settings"));
 	} 
 }
